apps/admin/pyproject.toml
[project]
name = "admin"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "grpcio>=1.62",
    "httpx>=0.27",
    "argon2-cffi>=23.1",
    "cryptography>=42.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
